automation_scripts/requirements.txt
adagio==0.2.6
antlr4-python3-runtime==4.11.1
appdirs==1.4.4
asn1crypto==1.5.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.3
datacompy==0.11.3
exceptiongroup==1.3.0
filelock==3.16.1
fs==2.4.16
fsspec==2025.3.0
fugue==0.8.7
fugue-sql-antlr==0.2.2
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==1.24.4
ordered-set==4.1.0
packaging==25.0
pandas==2.0.3
platformdirs==4.3.6
pluggy==1.5.0
pyarrow==17.0.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pyopenssl==25.1.0
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qpd==0.4.4
requests==2.32.3
six==1.17.0
snowflake-connector-python==3.14.0
snowflake-sqlalchemy==1.7.3
sortedcontainers==2.4.0
sqlalchemy==2.0.41
sqlglot==26.24.0
tomli==2.2.1
tomlkit==0.13.2
triad==0.9.8
typing-extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20